ocr_module/latex_to_image/requirements.txt
# LaTeX to Image 转换模块依赖包

# 基础依赖
numpy>=1.21.0
matplotlib>=3.3.0
Pillow>=8.3.0

# 开发依赖
pytest>=6.2.0